Added a characteristics.py file to house all the logic/conditions for solving a figure's value. The figures.py now imports the characteristics.py file and uses it as a class.
The Down, and Across subclasses now have a 'feature' attribute which will be used to determine the calculations of where their squares reside.

TODO: Need to alter the gameBoard to instead of being a size x size grid, just for it to be a list of integers... maybe. I'm still tweaking a few of the details, maybe a grid with an x-coordinate, and a y-coordinate might actually be good?