UPDATES TO GAME BOARD CLASS: a) The game board grid now will utilize a coordinate system using x and y values. b) It now successfully constructs a grid upon initialization of a GameBoard object when provided with a size parameter. c) I have created a classification system for cells using enums. (This was chosen in favour over utilizing multiple boolean attributes.) This has rendered a few methods in other classes worthless-- but will make readability that much better.
d) put a few methods on the back-burner until I sort through the logic of how the board will be populated, and figures will be related to the grid.

UPDATES TO THE FIGURES CLASS:
a) The Down, and Across subclasses have been eliminated in favour of a single boolean attribute: "is_across", by default this will be set to 'True'. This was done because I felt that Down and Across didn't have enough content to be their own subclasses.